Add files via upload
Table data description:
There are 112 artificial missing data have been identified from the 10 batches of both case cities temperature datasets as shown in table of data description of the Amsterdam’s and Bristol’s artificial missing data. The percentage of missing data in the dataset is 1.28%, which is according to the acceptance proportion proposed by Bennet in (Bennett 2001). The shortest sequence consists of 3 records and the longest sequence consists of 22 records.

Gambar Lost function:
epochs = 400  is assigned for the model training since the training loss graphic was saddled at that value.